Initial commit. This is still a WIP
Scalar multiplication is tested and working. Comb code is written and in the
process of being tested. It gives the correct result for the first everything
point. This means it is getting the field inversion correct. Have to make sure
that walking the gray code is correct and that comb multiplication is correct.