Simplify the algorithm for simultaneous square root and inversion
For p = 3 mod 4, you can compute sqrt(1/x) by x^((p-3)/4). So to compute
sqrt(x/y), it is sufficient to compute x(xy^((p-3)/4)). This follows
because x^((p+1)/4) is sqrt(x). The two exponents differ by exactly 1.

This adjusts the specific result that we get, but there is now an
additional check that the resulting value is in fact sqrt(x/y). Checked
by verifying that x = (sqrt(x/y)^2 * y).